+++ docs/realm.js
@@ -65,6 +65,14 @@ class Realm {
     get schemaVersion() {}
 
     /**
+     * Indicates if this Realm is in a write transaction.
+     * @type {boolean}
+     * @readonly
+     * @since 1.10.3
+     */
+    get isInTransaction() {}
+
+    /**
      * Gets the sync session if this is a synced Realm
      * @type {Session}
      */
@@ -182,6 +190,22 @@ class Realm {
     * @param {function()} callback
     */
     write(callback) {}
+
+    /**
+     * Initiate a write transaction.
+     * @throws {Error} When already in write transaction
+     */
+    beginTransaction() {}
+
+    /**
+     * Commit a write transaction.
+     */
+    commitTransaction() {}
+
+    /**
+     * Cancel a write transaction.
+     */
+    cancelTransaction() {}
 }
 
 /**